backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
aiosqlite>=0.20.0
